pyproject.toml
[tool.poetry]
name = "langchain-experiments"
version = "0.1.0"
description = ""
authors = ["Bharath Sriraam R R <albharath.1305@gmail.com>"]
readme = "README.md"
packages = [{include = "langchain-experiments"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.212"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
faiss-cpu = "^1.7.4"
tiktoken = "^0.4.0"
ipykernel = "^6.23.2"
chromadb = "^0.3.26"
lark = "^1.1.5"
yt-dlp = "^2023.6.22"
pydub = "^0.25.1"
replicate = "^0.8.3"
beautifulsoup4 = "^4.12.2"
fastapi = "^0.98.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
